refactor: extract duplicated find_project_root() into shared CLI util
`find_project_root()` was copy-pasted identically into both
`commands/clean.rs` and `commands/deps/cli.rs`. Move it to the existing
`commands/util.rs` shared helpers module and update both callers to
import from there instead.

Zero behaviour change — the function body is byte-for-byte identical
across all copies.

Closes #3183
